Add fastNoiseLambda
Essentially a 3D version of valueNoiseIQ.

The original function used lambda, because those prove really useful in the context of vertex shaders -- which i am exploring.

But Spark AR Studio does not allow it to be marked as an export. So I added a regular version for export, and added notes on how to turn it into a lambda function.

I also slightly modified shaderCodeAsset.sca to allow for more flexibility in the future.